+++ src/main/scala/org/podval/tools/scalajsplugin/nonjvm/NonJvmTask.scala
@@ -1,6 +1,6 @@
 package org.podval.tools.scalajsplugin.nonjvm
 
-import org.gradle.api.Task
+import org.podval.tools.scalajsplugin.BackendTask
 
-trait NonJvmTask[L <: NonJvmLinkTask[L]] extends Task:
+trait NonJvmTask[L <: NonJvmLinkTask[L]] extends BackendTask:
   protected def linkTask: L