+++ src/main/scala/org/podval/tools/scalajsplugin/nonjvm/NonJvmTestTask.scala
@@ -1,5 +1,0 @@
-package org.podval.tools.scalajsplugin.nonjvm
-
-import org.podval.tools.test.task.TestTask
-
-abstract class NonJvmTestTask[L <: NonJvmLinkTask[L]] extends TestTask with NonJvmRunTask[L]