+++ src/main/scala/org/podval/tools/scalajsplugin/scalajs/ScalaJSLinkTestTask.scala
@@ -1,7 +1,7 @@
 package org.podval.tools.scalajsplugin.scalajs
 
 import org.podval.tools.scalajs.ModuleInitializer
-import org.podval.tools.scalajsplugin.nonjvm.NonJvmLinkTestTask
+import org.podval.tools.scalajsplugin.nonjvm.NonJvmLinkTask
 
-abstract class ScalaJSLinkTestTask extends NonJvmLinkTestTask[ScalaJSLinkTask] with ScalaJSLinkTask:
+abstract class ScalaJSLinkTestTask extends NonJvmLinkTask.Test[ScalaJSLinkTask] with ScalaJSLinkTask:
   final override def moduleInitializers: Option[Seq[ModuleInitializer]] = None