+++ src/main/scala/org/podval/tools/scalajsplugin/scalajs/ScalaJSTestTask.scala
@@ -1,9 +1,9 @@
 package org.podval.tools.scalajsplugin.scalajs
 
 import org.podval.tools.build.ScalaBackendKind
-import org.podval.tools.scalajsplugin.nonjvm.NonJvmTestTask
+import org.podval.tools.scalajsplugin.nonjvm.NonJvmRunTask
 
-abstract class ScalaJSTestTask extends NonJvmTestTask[ScalaJSLinkTask] with ScalaJSRunTask:
+abstract class ScalaJSTestTask extends NonJvmRunTask.Test[ScalaJSLinkTask] with ScalaJSRunTask:
   final override protected def linkTaskClass: Class[ScalaJSLinkTestTask] = classOf[ScalaJSLinkTestTask]
 
   final override protected def backendKind: ScalaBackendKind = ScalaBackendKind.JS